+++ src/EndpointList.php
@@ -68,7 +68,6 @@ class EndpointList
 
   /**
    * Add endpoint
-   * @todo Prevent cross-service injection
    */
   public function add(Endpoint $endpoint): static
   {
@@ -78,7 +77,6 @@ class EndpointList
 
   /**
    * Remove endpoint
-   * @todo Prevent cross-service injection
    */
   public function remove(Endpoint|string $endpointOrPath): static
   {