+++ src/traits/HasRequestHooks.php
@@ -55,6 +55,7 @@ trait HasRequestHooks
    */
   public function hookBefore(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::Before, $handler);
     return $this;
@@ -67,6 +68,7 @@ trait HasRequestHooks
    */
   public function hookBeforeGet(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::BeforeGet, $handler);
     return $this;
@@ -79,6 +81,7 @@ trait HasRequestHooks
    */
   public function hookBeforePost(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::BeforePost, $handler);
     return $this;
@@ -91,6 +94,7 @@ trait HasRequestHooks
    */
   public function hookBeforeHead(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::BeforeHead, $handler);
     return $this;
@@ -103,6 +107,7 @@ trait HasRequestHooks
    */
   public function hookBeforePut(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::BeforePut, $handler);
     return $this;
@@ -115,6 +120,7 @@ trait HasRequestHooks
    */
   public function hookBeforePatch(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::BeforePatch, $handler);
     return $this;
@@ -127,6 +133,7 @@ trait HasRequestHooks
    */
   public function hookBeforeDelete(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::BeforeDelete, $handler);
     return $this;
@@ -139,6 +146,7 @@ trait HasRequestHooks
    */
   public function hookAfter(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::After, $handler);
     return $this;
@@ -151,6 +159,7 @@ trait HasRequestHooks
    */
   public function hookAfterGet(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::AfterGet, $handler);
     return $this;
@@ -163,6 +172,7 @@ trait HasRequestHooks
    */
   public function hookAfterPost(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::AfterPost, $handler);
     return $this;
@@ -175,6 +185,7 @@ trait HasRequestHooks
    */
   public function hookAfterHead(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::AfterHead, $handler);
     return $this;
@@ -187,6 +198,7 @@ trait HasRequestHooks
    */
   public function hookAfterPut(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::AfterPut, $handler);
     return $this;
@@ -199,6 +211,7 @@ trait HasRequestHooks
    */
   public function hookAfterPatch(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::AfterPatch, $handler);
     return $this;
@@ -211,6 +224,7 @@ trait HasRequestHooks
    */
   public function hookAfterDelete(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::AfterDelete, $handler);
     return $this;
@@ -223,6 +237,7 @@ trait HasRequestHooks
    */
   public function hookOnError(callable $handler): static
   {
+    $this->_assertNotLocked('add hook');
     $this->initRequestHooks();
     $this->hooks->add(RequestHookKey::OnError, $handler);
     return $this;